+++ CHANGELOG.md
@@ -98,6 +98,8 @@
   * Resolved issues with drift detection after deletion of key version in `akamai_cloudaccess_key` resource. ([I#579](https://github.com/akamai/terraform-provider-akamai/issues/579))
 
 
+* PAPI
+  * Fixed idempotency issue in property activation when `rule_errors` is empty.